gui: restructuring
