+++ xmodits-gui/src/core/dialog.rs
