+++ xmodits-gui/src/core/mod.rs
@@ -1,6 +1,8 @@
 pub mod cfg;
-pub mod font;
 pub mod log;
 pub mod sfx;
 pub mod xmodits;
+pub mod entries;
+pub mod dialog;
+pub mod extraction;
 // pub use sfx::SFX;
